[REFACTOR/#210] 광고 배너 컴포넌트 core:ui 모듈로 이전
