+++ core/ui/build.gradle.kts
@@ -1,4 +1,5 @@
 import com.yapp.convention.setNamespace
+import java.util.Properties
 
 plugins {
     id("orbit.android.library")
@@ -7,6 +8,28 @@ plugins {
 
 android {
     setNamespace("core.ui")
+
+    buildTypes {
+        val localProperties = Properties()
+        localProperties.load(
+            project.rootProject.file("local.properties").bufferedReader(),
+        )
+
+        debug {
+            resValue(
+                "string",
+                "admob_ad_unit_id",
+                localProperties["admobAdUnitIdDebug"] as String,
+            )
+        }
+        release {
+            resValue(
+                "string",
+                "admob_ad_unit_id",
+                localProperties["admobAdUnitIdRelease"] as String,
+            )
+        }
+    }
 }
 
 dependencies {
@@ -17,4 +40,5 @@ dependencies {
     implementation(libs.orbit.compose)
     implementation(libs.orbit.viewmodel)
     implementation(libs.lottie.compose)
+    implementation(libs.play.services.ads)
 }
