+++ core/ui/src/main/java/com/yapp/ui/component/banner/AdsBanner.kt
@@ -1,0 +1,34 @@
+package com.yapp.ui.component.banner
+
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.viewinterop.AndroidView
+import com.google.android.gms.ads.AdRequest
+import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.AdView
+import core.ui.R
+
+@SuppressLint("MissingPermission")
+@Composable
+fun AdsBanner(
+    modifier: Modifier = Modifier,
+) {
+    val adUnitId = stringResource(id = R.string.admob_ad_unit_id)
+
+    AndroidView(
+        modifier = modifier.fillMaxWidth(),
+        factory = { context ->
+            AdView(context).apply {
+                setAdSize(AdSize.BANNER)
+                this.adUnitId = adUnitId
+                loadAd(AdRequest.Builder().build())
+            }
+        },
+        update = { adView ->
+            adView.loadAd(AdRequest.Builder().build())
+        },
+    )
+}
